+++ react/IconVolumeOff.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import Icon from './components/Icon'
 
 const IconVolumeOff = (props: IconProps) => {
-  return <Icon id="sti-volume-off" handle="volumeOffIcon" {...props} />
+  return <Icon {...props} id="sti-volume-off" handle="volumeOffIcon" />
 }
 
 export default IconVolumeOff