+++ react/__mocks__/vtex.css-handles.ts
@@ -1,3 +1,5 @@
 export function useCssHandles() {
-  return { handle: 'handle' }
+  return new Proxy({}, {
+    get: () => '',
+  })
 }