+++ react/typings/vtex.css-handles.d.ts
@@ -1,0 +1,5 @@
+declare module 'vtex.css-handles' {
+  export function useCssHandles(handles: readonly string[]): {
+    [key: string]: string
+  }
+}